requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
plotly==5.24.1
openpyxl==3.1.5
python-dateutil==2.9.0.post0
pytest==8.3.2
reportlab==4.2.2
kaleido==0.2.1
openai>=1.40.0
